+++ test/spec/user-container.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var UserContainer = require('../lib/authentication/user-container');
+var UserContainer = require('../../lib/authentication/user-container');
 var assert = require('chai').assert;
 
 describe('tests of garbage collecting of outdated cookies', function() {
